u5w7d2pratica

+++ src/main/java/backend_esame4/Dispositivi_Aziendali/repository/LaptopRepository.java
@@ -2,5 +2,5 @@ package backend_esame4.Dispositivi_Aziendali.repository;
 import backend_esame4.Dispositivi_Aziendali.model.Laptop;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LaptopRepository extends JpaRepository<Laptop, Long>{
+public interface LaptopRepository extends JpaRepository<Laptop, Integer>{
 }
