+++ src/main/java/backend_esame4/Dispositivi_Aziendali/repository/SmartphoneRepository.java
@@ -2,5 +2,5 @@ package backend_esame4.Dispositivi_Aziendali.repository;
 import backend_esame4.Dispositivi_Aziendali.model.Smartphone;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SmartphoneRepository extends JpaRepository<Smartphone, Long> {
+public interface SmartphoneRepository extends JpaRepository<Smartphone, Integer> {
 }
