+++ src/main/java/backend_esame4/Dispositivi_Aziendali/repository/TabletRepository.java
@@ -3,5 +3,5 @@ import backend_esame4.Dispositivi_Aziendali.model.Laptop;
 import backend_esame4.Dispositivi_Aziendali.model.Tablet;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TabletRepository extends JpaRepository<Tablet, Long> {
+public interface TabletRepository extends JpaRepository<Tablet, Integer> {
 }
